+++ src/Serilog.Sinks.Http/Sinks/Http/IHttpClient.cs
@@ -13,15 +13,19 @@
 // limitations under the License.
 
 using System;
+using System.IO;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
+using Serilog.Sinks.Http.HttpClients;
 
 namespace Serilog.Sinks.Http
 {
     /// <summary>
     /// Interface responsible for posting HTTP requests.
     /// </summary>
+    /// <seealso cref="JsonHttpClient"/>
+    /// <seealso cref="JsonGzipHttpClient"/>
     public interface IHttpClient : IDisposable
     {
         /// <summary>
@@ -34,8 +38,8 @@ namespace Serilog.Sinks.Http
         /// Sends a POST request to the specified Uri as an asynchronous operation.
         /// </summary>
         /// <param name="requestUri">The Uri the request is sent to.</param>
-        /// <param name="content">The HTTP request content sent to the server.</param>
+        /// <param name="contentStream">The stream containing the content of the request.</param>
         /// <returns>The task object representing the asynchronous operation.</returns>
-        Task<HttpResponseMessage> PostAsync(string requestUri, HttpContent content);
+        Task<HttpResponseMessage> PostAsync(string requestUri, Stream contentStream);
     }
 }